1，解决了头像的上传以及在各处的显示的问题。 2，解决了在联系人界面下收到好友请求，新好友的角标不显示的问题。

+++ ComChat/ViewsController/FriendManageViewController.m
@@ -167,11 +167,11 @@
 
 
 
-- (float)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section {
+- (CGFloat)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section {
     return 20;
 }
 
-- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     return 45.0f;
 }
